fix lesson8 task 2

+++ Lesson8/src/main/java/by/tms/task2/model/RequiredFieldsCarClassesAware.java
@@ -6,7 +6,7 @@ public interface RequiredFieldsCarClassesAware {
 
     void startCar();
 
-    void go(int distanceTraveled);
+    void go(double distanceTraveled);
 
     void stopCar();
 }
